refactor : 커스텀 훅 분리 #92
